# openapi.yaml
openapi: 3.0.4
info:
  title: 'PhotoBank.Api, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null'
  version: '1.0'
servers:
  - url: /api
paths:
  /admin/access-profiles:
    get:
      tags:
        - AdminAccessProfiles
      operationId: AdminAccessProfiles_List
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AccessProfile'
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AccessProfile'
            text/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AccessProfile'
    post:
      tags:
        - AdminAccessProfiles
      operationId: AdminAccessProfiles_Create
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccessProfile'
          text/json:
            schema:
              $ref: '#/components/schemas/AccessProfile'
          application/*+json:
            schema:
              $ref: '#/components/schemas/AccessProfile'
      responses:
        '200':
          description: OK
  '/admin/access-profiles/{id}':
    get:
      tags:
        - AdminAccessProfiles
      operationId: AdminAccessProfiles_Get
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/AccessProfile'
            application/json:
              schema:
                $ref: '#/components/schemas/AccessProfile'
            text/json:
              schema:
                $ref: '#/components/schemas/AccessProfile'
    put:
      tags:
        - AdminAccessProfiles
      operationId: AdminAccessProfiles_Update
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccessProfile'
          text/json:
            schema:
              $ref: '#/components/schemas/AccessProfile'
          application/*+json:
            schema:
              $ref: '#/components/schemas/AccessProfile'
      responses:
        '200':
          description: OK
    delete:
      tags:
        - AdminAccessProfiles
      operationId: AdminAccessProfiles_Delete
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: OK
  '/admin/access-profiles/{id}/assign-user/{userId}':
    post:
      tags:
        - AdminAccessProfiles
      operationId: AdminAccessProfiles_AssignUser
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
    delete:
      tags:
        - AdminAccessProfiles
      operationId: AdminAccessProfiles_UnassignUser
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
  '/admin/access-profiles/{id}/assign-role/{roleId}':
    post:
      tags:
        - AdminAccessProfiles
      operationId: AdminAccessProfiles_AssignRole
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
        - name: roleId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
...
  '/admin/users/{id}':
    put:
      tags:
        - Users
      operationId: Users_Update
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserDto'
          text/json:
            schema:
              $ref: '#/components/schemas/UpdateUserDto'
          application/*+json:
            schema:
              $ref: '#/components/schemas/UpdateUserDto'
      responses:
        '200':
          description: OK
        '404':
          description: Not Found
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '400':
          description: Bad Request
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  '/admin/users/{id}/claims':
    put:
      tags:
        - Users
      operationId: Users_SetClaims
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/ClaimDto'
          text/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/ClaimDto'
          application/*+json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/ClaimDto'
      responses:
        '200':
          description: OK
        '404':
          description: Not Found
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '400':
          description: Bad Request
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
components:
  schemas:
    AccessProfile:
      type: object
      properties:
        id:
          type: integer
          format: int32
        name:
          maxLength: 128
          type: string
          nullable: true
        description:
          maxLength: 512
          type: string
          nullable: true
        flags_CanSeeNsfw:
          type: boolean
        storages:
          type: array
          items:
            $ref: '#/components/schemas/AccessProfileStorageAllow'
          nullable: true
        personGroups:
          type: array
          items:
            $ref: '#/components/schemas/AccessProfilePersonGroupAllow'
          nullable: true
        dateRanges:
          type: array
          items:
            $ref: '#/components/schemas/AccessProfileDateRangeAllow'
          nullable: true
      additionalProperties: false
    AccessProfileDateRangeAllow:
      type: object
      properties:
        profileId:
          type: integer
          format: int32
        fromDate:
          type: string
          format: date
        toDate:
          type: string
          format: date
        profile:
          $ref: '#/components/schemas/AccessProfile'
      additionalProperties: false
    AccessProfilePersonGroupAllow:
      type: object
      properties:
        profileId:
          type: integer
          format: int32
        personGroupId:
          type: integer
          format: int32
        profile:
          $ref: '#/components/schemas/AccessProfile'
      additionalProperties: false
    AccessProfileStorageAllow:
      type: object
      properties:
        profileId:
          type: integer
          format: int32
        storageId:
          type: integer
          format: int32
        profile:
          $ref: '#/components/schemas/AccessProfile'
      additionalProperties: false
    ClaimDto:
      required:
        - type
        - value
      type: object
      properties:
        type:
          type: string
          nullable: true
        value:
          type: string
          nullable: true
      additionalProperties: false
...
    ProblemDetails:
      type: object
      properties:
        type:
          type: string
          nullable: true
        title:
          type: string
          nullable: true
        status:
          type: integer
          format: int32
          nullable: true
        detail:
          type: string
          nullable: true
        instance:
          type: string
          nullable: true
      additionalProperties: { }
...
    UpdateUserDto:
      type: object
      properties:
        phoneNumber:
          type: string
          nullable: true
        telegramUserId:
          type: integer
          format: int64
          nullable: true
        telegramSendTimeUtc:
          type: string
          format: date-span
          nullable: true
      additionalProperties: false
...